app/requirements.txt
asgiref==3.11.0
attrs==25.4.0
Django==5.2.8
django-environ==0.12.0
django-jazzmin==3.0.1
djangorestframework==3.14.0
drf-spectacular==0.29.0
drf-spectacular-sidecar==2025.10.1
gunicorn==23.0.0
inflection==0.5.1
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
packaging==25.0
pillow==12.0.0
psycopg2==2.9.11
pytz==2025.2
PyYAML==6.0.3
qrcode==8.2
referencing==0.37.0
rpds-py==0.29.0
sqlparse==0.5.4
typing_extensions==4.15.0
uritemplate==4.2.0
